modul2_weather_ops/requirements-core.txt
pandas
pydantic
requests
plotly
python-dotenv
pyyaml
